create separate TableExtension
