+++ packages/core/src/BlockNoteEditor.ts
@@ -35,7 +35,6 @@ import {
 import { Selection } from "./extensions/Blocks/api/selectionTypes";
 import { getBlockInfoFromPos } from "./extensions/Blocks/helpers/getBlockInfoFromPos";
 
-import { columnResizing, tableEditing } from "prosemirror-tables";
 import "prosemirror-tables/style/tables.css";
 import "./editor.css";
 import { FormattingToolbarProsemirrorPlugin } from "./extensions/FormattingToolbar/FormattingToolbarPlugin";
@@ -208,10 +207,6 @@ export class BlockNoteEditor<BSchema extends BlockSchema = DefaultBlockSchema> {
           this.hyperlinkToolbar.plugin,
           this.imageToolbar.plugin,
           this.tableHandles.plugin,
-          columnResizing({
-            cellMinWidth: 100,
-          }),
-          tableEditing(),
         ];
       },
     });
